Define Segment class
Implemented a class called "Segment" to handle loading segment
reconstructions and reading their corresponding headers.

The scan is loaded into a 3D numpy array. From the header file, the
class reads information on scan dimensions, voxel size and bed position.
Additionally, the class calculates the z-positions of all transversal
slices. Later on, the z-positions and the horizontal bed position will
be used to align the scans.