Test whether scan properties match
Implemented a method that returns "True" if certain scan properties of
the upper and lower segment match. These properties are scan dimensions,
voxel sizes, and vertical bed positions.

To ease comparability, can dimensions and voxel sizes are not stored as
dictionaries any more. Instead they are stored as numpy arrays. If their
Euclidean distance is 0, the match.